fixed bugs

+++ server/schemas/resolvers.js
@@ -1,18 +1,20 @@
 const { AuthenticationError } = require('apollo-server-express');
-const { User, Book } = require('../models');
+const { User } = require('../models');
 const { signToken } = require('../utils/auth');
 
 const resolvers = {
     Query: {
-        me: async (parent, arg, context) => {if (context.user) {
-            return User.findOne({ _id: context.user._id }).populate('thoughts');
-          }
-          throw new AuthenticationError('You need to be logged in!');
+        me: async (parent, arg, context) => {
+            if (context.user) {
+                const userData = User.findOne({ _id: context.user._id }).populate('books');
+                return userData;
+            }
+            throw new AuthenticationError('You need to be logged in!');
         }
     },
     Mutation: {
         login: async (parent, { email, password }) => {
-            const user = User.findOne( { email: email } );
+            const user = User.findOne({ email: email });
             if (!user) {
                 throw new AuthenticationError('No profile with this email found!');
             }
@@ -21,7 +23,7 @@ const resolvers = {
                 throw new AuthenticationError('Wrong password');
             }
             const token = signToken(user);
-            return ({ token, user });
+            return { token, user };
         },
         addUser: async (parent, { username, email, password }) => {
             const user = await User.create({
@@ -32,30 +34,29 @@ const resolvers = {
             const token = signToken(user);
             return { token, user };
         },
-        saveBook: async (parent, { book }, context) => {
+        saveBook: async (_parent, { userId, bookData }, context) => {
             if (context.user) {
-                const book = await Book.create({
-                    description: book.description,
-                    title: book.title,
-                });
-            }
-            await User.findOneAndUpdate(
-                { _id: context.user.id },
-                { $addToSet: { savedBooks: book.id } },
-                { new: true, runValidators: true }
-            );
-            return book;
-        },
-        removeBook: async (parent, {bookId, userId}) => {
-            await Book.findOneAndDelete({
-                _id: bookId,
-            })
-            const user = await User.findOneAndUpdate(
+              return User.findOneAndUpdate(
                 { _id: userId },
-                { $pull: { savedBooks: { bookId: bookId } } },
-                { new: true }
-            );
-            return user;
+                { $addToSet: { savedBooks: { book: bookData } } },
+                { new: true, runValidators: true }
+              );
+            }
+            throw new AuthenticationError("Please login or register");
+          },
+        removeBook: async (parent, { bookId }, context) => {
+            if (context.user) {
+                await Book.findOneAndDelete({
+                    _id: bookId,
+                })
+                const user = await User.findOneAndUpdate(
+                    { _id: context.user._id },
+                    { $pull: { savedBooks: { bookId: bookId } } },
+                    { new: true }
+                );
+                return user;
+            }
+            throw new AuthenticationError('Wrong password');
         },
     }
 }
